+++ src/components/App.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import './App.css';
+import '../App.css';
 import Cards from './Cards'
 import Bookshelf from './Bookshelf'
